Squashed commit of the following:
commit 697a58924c835bf10de11c5741bccc19d5142a90
Author: hyun0404woo <hyun0404woo@gmail.com>
Date:   Wed Nov 9 20:19:11 2022 +0900

    chore: 패키지 구조 수정

    - repository 패키지 명시
    - 계층형 구조로 유지
      - 추후 프로젝트가 커지면 도메인형으로 변경 예정
