+++ src/main/java/org/goormton/darktourism/controller/member/MemberControllerImpl.java
@@ -7,9 +7,9 @@ import org.goormton.darktourism.controller.badge.mapper.SimpleBadgeDtoMapper;
 import org.goormton.darktourism.controller.member.dto.MemberInfoAll;
 import org.goormton.darktourism.controller.place.dto.SimplePlaceDto;
 import org.goormton.darktourism.controller.place.mapper.SimplePlaceDtoMapper;
-import org.goormton.darktourism.domain.Badge;
-import org.goormton.darktourism.domain.Member;
-import org.goormton.darktourism.domain.Place;
+import org.goormton.darktourism.domain.badge.Badge;
+import org.goormton.darktourism.domain.member.Member;
+import org.goormton.darktourism.domain.place.Place;
 import org.goormton.darktourism.service.badge.BadgeService;
 import org.goormton.darktourism.service.member.MemberService;
 import org.goormton.darktourism.service.place.PlaceService;
