+++ src/main/java/org/goormton/darktourism/controller/place/PlaceControllerImpl.java
@@ -4,22 +4,17 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.goormton.darktourism.controller.auth.dto.LoginDto;
 import org.goormton.darktourism.controller.place.dto.SimplePlaceDto;
-import org.goormton.darktourism.controller.place.dto.VisitPlaceDto;
 import org.goormton.darktourism.controller.place.mapper.PlaceToDtoMapper;
 import org.goormton.darktourism.controller.place.mapper.SimplePlaceDtoMapper;
-import org.goormton.darktourism.domain.Member;
-import org.goormton.darktourism.domain.Place;
+import org.goormton.darktourism.domain.member.Member;
+import org.goormton.darktourism.domain.place.Place;
 import org.goormton.darktourism.service.badge.BadgeService;
 import org.goormton.darktourism.service.member.MemberService;
 import org.goormton.darktourism.service.place.PlaceService;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.validation.Valid;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
