+++ src/main/java/org/goormton/darktourism/domain/badge/BadgeMember.java
@@ -1,9 +1,10 @@
-package org.goormton.darktourism.domain;
+package org.goormton.darktourism.domain.badge;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.goormton.darktourism.domain.member.Member;
 
 import javax.persistence.*;
 
@@ -12,31 +13,28 @@ import static javax.persistence.GenerationType.IDENTITY;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class PlaceStarMember {
-
+public class BadgeMember {
     @Id
     @GeneratedValue(strategy = IDENTITY)
+    @Column(name = "badge_number_id")
     private Long id;
-
-    private Double star = 0.0;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "place_id")
-    private Place place;
+    @JoinColumn(name = "badge_id")
+    private Badge badge;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
     private Member member;
 
-    private PlaceStarMember(Place place, Member member) {
-        this.place = place;
+    private BadgeMember(Badge badge, Member member) {
+        this.badge = badge;
         this.member = member;
-        this.star = 0.0;
     }
 
-    public static PlaceStarMember createPlaceStarMember(Place place, Member member) {
-        return new PlaceStarMember(place, member);
+    public static BadgeMember createBadgeMember(Badge badge, Member member) {
+        return new BadgeMember(badge, member);
     }
 }
