--- src/main/java/org/goormton/darktourism/exception/ForbiddenException.java
@@ -1,13 +1,0 @@
-package org.goormton.darktourism.exception;
-
-import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.FORBIDDEN)
-@NoArgsConstructor
-public class ForbiddenException extends RuntimeException {
-    public ForbiddenException(String message) {
-        super(message);
-    }
-}
