+++ src/main/java/org/goormton/darktourism/repository/badge/BadgeMemberRepository.java
@@ -1,8 +1,7 @@
-package org.goormton.darktourism.repository;
+package org.goormton.darktourism.repository.badge;
 
-import org.goormton.darktourism.domain.BadgeMember;
-import org.goormton.darktourism.domain.Member;
-import org.goormton.darktourism.domain.PlaceStarMember;
+import org.goormton.darktourism.domain.badge.BadgeMember;
+import org.goormton.darktourism.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
