+++ src/main/java/org/goormton/darktourism/repository/member/MemberRepository.java
@@ -1,6 +1,6 @@
-package org.goormton.darktourism.repository;
+package org.goormton.darktourism.repository.member;
 
-import org.goormton.darktourism.domain.Member;
+import org.goormton.darktourism.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
