+++ src/main/java/org/goormton/darktourism/repository/place/PlaceImageUrlRepository.java
@@ -1,7 +1,7 @@
-package org.goormton.darktourism.repository;
+package org.goormton.darktourism.repository.place;
 
 
-import org.goormton.darktourism.domain.PlaceImageUrl;
+import org.goormton.darktourism.domain.place.PlaceImageUrl;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
