+++ src/main/java/org/goormton/darktourism/repository/place/PlaceRepository.java
@@ -1,6 +1,6 @@
-package org.goormton.darktourism.repository;
+package org.goormton.darktourism.repository.place;
 
-import org.goormton.darktourism.domain.Place;
+import org.goormton.darktourism.domain.place.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
