+++ src/main/java/org/goormton/darktourism/repository/place/PlaceStarMemberRepository.java
@@ -1,8 +1,8 @@
-package org.goormton.darktourism.repository;
+package org.goormton.darktourism.repository.place;
 
-import org.goormton.darktourism.domain.Member;
-import org.goormton.darktourism.domain.Place;
-import org.goormton.darktourism.domain.PlaceStarMember;
+import org.goormton.darktourism.domain.member.Member;
+import org.goormton.darktourism.domain.place.Place;
+import org.goormton.darktourism.domain.place.PlaceStarMember;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
